refactor(infrastructure): split file_config.rs into submodule directory (#222)
Split the 1773-line monolithic file_config.rs into 9 focused submodules
organized by TOML section, following the same directory pattern as
domain/agent/ and domain/tool/.
