+++ infrastructure/src/config/file_config/output.rs
@@ -1,0 +1,41 @@
+//! Output configuration from TOML (`[output]` section)
+
+use quorum_domain::OutputFormat;
+use serde::{Deserialize, Serialize};
+
+// Re-export OutputFormat from domain for convenience
+pub use quorum_domain::OutputFormat as FileOutputFormat;
+
+/// Raw output configuration from TOML
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+pub struct FileOutputConfig {
+    /// Output format (uses domain type)
+    pub format: Option<OutputFormat>,
+    /// Enable colored terminal output
+    pub color: bool,
+}
+
+impl Default for FileOutputConfig {
+    fn default() -> Self {
+        Self {
+            format: None,
+            color: true,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_output_format_deserialize() {
+        let toml_str = r#"
+[output]
+format = "json"
+"#;
+        let config: super::super::FileConfig = toml::from_str(toml_str).unwrap();
+        assert_eq!(config.output.format, Some(OutputFormat::Json));
+    }
+}
