+++ infrastructure/src/config/file_config/repl.rs
@@ -1,0 +1,22 @@
+//! REPL configuration from TOML (`[repl]` section)
+
+use serde::{Deserialize, Serialize};
+
+/// Raw REPL configuration from TOML
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+pub struct FileReplConfig {
+    /// Show progress indicators
+    pub show_progress: bool,
+    /// Path to history file
+    pub history_file: Option<String>,
+}
+
+impl Default for FileReplConfig {
+    fn default() -> Self {
+        Self {
+            show_progress: true,
+            history_file: None,
+        }
+    }
+}
